server/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.7.1
pymysql==1.1.1
bcrypt==3.2.0
email-validator==2.2.0
httpx==0.28.1
ultralytics==8.3.70
opencv-python==4.10.0.84
python-dotenv==1.0.1
google-auth==2.38.0
fastapi-mail==1.4.1
alembic==1.14.1
